+++ lib/src/chart/bar_chart/bar_chart_data.dart
@@ -683,6 +683,9 @@ class BarTouchTooltipData with EquatableMixin {
   /// Controls the rotation of the tooltip.
   final double rotateAngle;
 
+  /// The tooltip border color.
+  final BorderSide tooltipBorder;
+
   /// if [BarTouchData.handleBuiltInTouches] is true,
   /// [BarChart] shows a tooltip popup on top of rods automatically when touch happens,
   /// otherwise you can show it manually using [BarChartGroupData.showingTooltipIndicators].
@@ -707,6 +710,7 @@ class BarTouchTooltipData with EquatableMixin {
     bool? fitInsideVertically,
     TooltipDirection? direction,
     double? rotateAngle,
+    BorderSide? tooltipBorder,
   })  : tooltipBgColor = tooltipBgColor ?? Colors.blueGrey.darken(15),
         tooltipRoundedRadius = tooltipRoundedRadius ?? 4,
         tooltipPadding = tooltipPadding ??
@@ -718,6 +722,7 @@ class BarTouchTooltipData with EquatableMixin {
         fitInsideVertically = fitInsideVertically ?? false,
         direction = direction ?? TooltipDirection.auto,
         rotateAngle = rotateAngle ?? 0.0,
+        tooltipBorder = tooltipBorder ?? BorderSide.none,
         super();
 
   /// Used for equality check, see [EquatableMixin].
@@ -732,6 +737,7 @@ class BarTouchTooltipData with EquatableMixin {
         fitInsideHorizontally,
         fitInsideVertically,
         rotateAngle,
+        tooltipBorder,
       ];
 }