+++ lib/src/chart/line_chart/line_chart_painter.dart
@@ -23,7 +23,8 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
       _extraLinesPaint,
       _touchLinePaint,
       _bgTouchTooltipPaint,
-      _imagePaint;
+      _imagePaint,
+      _borderTouchTooltipPaint;
 
   /// Paints [dataList] into canvas, it is the animating [LineChartData],
   /// [targetData] is the animation's target and remains the same
@@ -56,6 +57,11 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
       ..color = Colors.white;
 
     _imagePaint = Paint();
+
+    _borderTouchTooltipPaint = Paint()
+      ..style = PaintingStyle.stroke
+      ..color = Colors.transparent
+      ..strokeWidth = 1.0;
   }
 
   /// Paints [LineChartData] into the provided canvas.
@@ -1136,6 +1142,11 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
     final textRotationOffset =
         Utils().calculateRotationOffset(rect.size, rotateAngle);
 
+    if (tooltipData.tooltipBorder != BorderSide.none) {
+      _borderTouchTooltipPaint.color = tooltipData.tooltipBorder.color;
+      _borderTouchTooltipPaint.strokeWidth = tooltipData.tooltipBorder.width;
+    }
+
     canvasWrapper.drawRotated(
       size: rect.size,
       rotationOffset: rectRotationOffset,
@@ -1143,6 +1154,7 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
       angle: rotateAngle,
       drawCallback: () {
         canvasWrapper.drawRRect(roundedRect, _bgTouchTooltipPaint);
+        canvasWrapper.drawRRect(roundedRect, _borderTouchTooltipPaint);
       },
     );